Configure Component Governance to ignore build\v8build\depot_tools (#93)
We have a lot of Component Governance (CG) alerts for this repo right now. Most of these appear to be coming from V8 code, which is downloaded at build time. This commit overwrites the auto-injected CG ADO task and adds a manual one that's configured to ignore build\v8build\depot_tools.