Patch V8 to not throw on attempt to freeze host objects (#119)
* React Native bridge tries to freeze all the object arguments when passed to the native side.
We have troubles with react native blobs because the blob object contains host objects in it (blob collector host object which cleans us the native blob object when javascript object is collected.)
This is because tryiing to freeze (Object.freeze) v8 host objects throws exception. This exception is not thrown when freezing Hermes host objects, very likely is silently failing.
It turns of V8 internally rightly fails the attemtp to freeze objects when it has interceptors, as V8 can't control external interceptors from mutating.
This patch suppresses the check. It is not correct to do it, but the effect will be similar to the current Hermes behaviour to silently fail the freeze attempt.

* fix: bumped package version, fixed build machine

Co-authored-by: shivenmian <shivenmian@microsoft.com>