[beta] Fixes for 12.2 (#189)
Populate the beta branch with the Beta version of V8 (12.2). This version included additional checks for isolate entrance/exit consistency which we were failing before, because apparently we kept the Isolated "entered" all the time on the thread.

This change implies that we may create more than one Isolate (even on the same thread) so that's something to look out for (potential perf impact).